requirements.txt
python-dotenv
pydantic==2.8.2
langchain==0.2.10
langchain-core==0.2.22
langchain_community==0.2.9
transformers==4.42.4
qdrant-client
langchain_qdrant
langchain_huggingface
pytest
numpy
scipy
unstructured[all-docs]  # This includes dependencies for all document types
pdfplumber
pandas
PyPDF2
tabula-py
torch==2.3.1+cu121
torchvision==0.18.1+cu121
lxml  # Required by unstructured
langchainhub
fuzzywuzzy
seaborn